Add Trace Response Handler To Server
